backend/requirements.txt
Flask
Flask-CORS
spacy
openai
python-dotenv
